feat: show install progress under instance card, no longer use dialog
