+++ src/pages/game/InstanceCard.vue
@@ -1,8 +1,6 @@
 <template>
   <div class="instance-card" :class="expand ? '' : 'instance-card-unexpand'" :style="banner">
-    <div
-      v-if="expand"
-      style="
+    <div v-if="expand" style="
         position: absolute;
         width: 100%;
         height: 100%;
@@ -23,14 +21,8 @@
         <div class="mod-loader-version" v-if="modLoaderType && modLoaderVersion">
           <img src="@/assets/images/fabric.webp" fill="#fff" v-if="modLoaderType === 'Fabric'" />
           <img src="@/assets/images/quilt.svg" fill="#fff" v-if="modLoaderType === 'Quilt'" />
-          <img
-            src="@/assets/images/neoforged.png"
-            fill="#fff"
-            v-if="modLoaderType === 'Neoforge'" />
-          <img
-            src="@/assets/images/Anvil_JE3_BE3.webp"
-            fill="#fff"
-            v-if="modLoaderType === 'Forge'" />
+          <img src="@/assets/images/neoforged.png" fill="#fff" v-if="modLoaderType === 'Neoforge'" />
+          <img src="@/assets/images/Anvil_JE3_BE3.webp" fill="#fff" v-if="modLoaderType === 'Forge'" />
           {{ modLoaderType }} {{ modLoaderVersion }}
         </div>
       </div>
@@ -47,20 +39,13 @@
         <i class="button gear"></i>
         <i class="button circle-info"></i>
         <i class="button star" id="star" @click="star"></i>
-        <button
-          class="game-button"
-          :class="`${gameButtonType}-game-button`"
-          @click="$emit('game-button-click')"
+        <button class="game-button" :class="`${gameButtonType}-game-button`" @click="$emit(gameButtonType)"
           v-if="!buttonLoading">
-          <i
-            :class="props.gameButtonType"
-            style="font-family: fa-pro; font-style: normal; margin-right: 5px; font-weight: 100"></i
-          >{{ gameButtonText }}
+          <i :class="gameButtonType"
+            style="font-family: fa-pro; font-style: normal; margin-right: 5px; font-weight: 100"></i>{{ gameButtonText
+          }}
         </button>
-        <button
-          class="game-button loading"
-          :class="`${gameButtonType}-game-button`"
-          v-if="buttonLoading">
+        <button class="game-button loading" :class="`${gameButtonType}-game-button`" v-if="buttonLoading">
           <item-loading-icon status="in-progress"></item-loading-icon>
         </button>
       </div>
@@ -82,7 +67,6 @@ const config = useConfigStore();
 const expand = ref(true);
 
 const props = defineProps<{
-  gameButtonType: "install" | "launch" | "error";
   buttonLoading: boolean;
   errorType?: "install" | "launch";
 }>();
@@ -103,6 +87,14 @@ const modLoaderVersion = computed(() => {
   return instanceStore.currentInstance.config.runtime.mod_loader_version;
 });
 
+const gameButtonType = computed(() => {
+  if (instanceStore.currentInstance.installed) {
+    return "launch";
+  } else {
+    return "install";
+  }
+});
+
 let computedInstanceName = computed(() => {
   let name = currentInstance.value.config.name;
   if (name == "Latest Release") {
@@ -116,7 +108,7 @@ let computedInstanceName = computed(() => {
 
 let banner = "";
 let gameButtonText = computed(() => {
-  switch (props.gameButtonType) {
+  switch (gameButtonType.value) {
     case "install":
       return i18n.t("game.install");
     case "launch":
