+++ src/pages/game/InstanceDetails.vue
@@ -1,23 +1,16 @@
 <template>
-  <div class="assets">
-    <tabs
-      :tabs="
-        components.map((n) => {
-          return i18n.t(n.name);
-        })
-      "
-      :active="activeTab"
-      :icons="
-        components.map((n) => {
-          return n.icon;
-        })
-      "
-      @choose-tab="chooseTab">
+  <div class="instance-details">
+    <tabs v-if="instanceStore.currentInstance.installed" :tabs="components.map((n) => {
+      return i18n.t(n.name);
+    })
+      " :active="activeTab" :icons="components.map((n) => {
+        return n.icon;
+      })
+        " @choose-tab="chooseTab">
     </tabs>
+    <tabs v-else :tabs="['Install Progress']" :icons="['folder-arrow-down']" :active="0"> </tabs>
     <Transition :name="transitionName" mode="out-in">
-      <component
-        :is="currentComponent"
-        style="padding: 16px 8px; width: 100%; height: fit-content"
+      <component :is="currentComponent" style="padding: 16px 8px; width: 100%; height: fit-content"
         @update-instance-list="$emit('update-instance-list')"></component>
     </Transition>
     <!-- <worlds :show="show.worlds" :datas="saves" :instance-name="instance.config.name" @close="show.worlds = false"> -->
@@ -39,9 +32,12 @@ import Worlds from "./Worlds.vue";
 import Mods from "./Mods.vue";
 import Packs from "./Packs.vue";
 import Settings from "./Settings.vue";
+import InstallProgress from "./InstallProgress.vue";
 import { useI18n } from "vue-i18n";
+import { useInstanceStore } from "@/store/instance";
 
 const i18n = useI18n();
+const instanceStore = useInstanceStore();
 
 defineEmits(["update-instance-list"]);
 
@@ -74,6 +70,8 @@ const components = ref([
 ]);
 const activeTab = ref(0);
 
+const transitionName = ref("slide-left");
+
 function chooseTab(tab: number) {
   if (activeTab.value < tab) {
     transitionName.value = "slide-left";
@@ -83,29 +81,31 @@ function chooseTab(tab: number) {
   activeTab.value = tab;
 }
 const currentComponent = computed(() => {
-  return components.value[activeTab.value].component;
+  if (instanceStore.currentInstance.installed) {
+    return components.value[activeTab.value].component;
+  } else {
+    return markRaw(InstallProgress);
+  }
 });
-
-const transitionName = ref("slide-left");
 </script>
 
 <style lang="less" scoped>
-.assets {
+.instance-details {
   margin-top: 14px;
   width: 100%;
   overflow-x: hidden;
-}
 
-.assets > div {
-  display: flex;
-  width: 100%;
-}
+  >div {
+    display: flex;
+    width: 100%;
+  }
 
-.assets > div.first-row {
-  margin-right: 5px;
-}
+  >div.first-row {
+    margin-right: 5px;
+  }
 
-.assets > div.second-row {
-  margin-left: 5px;
+  >div.second-row {
+    margin-left: 5px;
+  }
 }
 </style>
